modularize step 4 - move pasties out and load them dynamically
